pyproject.toml
[tool.poetry]
name = "django-site"
version = "0.1.0"
description = ""
authors = ["alexey <a556077@icloud.com>"]
readme = "README.md"
packages = [{include = "django_site"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
minio = "^7.1.17"
django-minio-storage = "^0.5.7"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.12"
django-redis = "^5.4.0"
psycopg-binary = "^3.1.13"
django-countries = "^7.5.1"
celery = {extras = ["redis"], version = "^5.3.6"}
django-filter = "^23.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
force_single_line = "True"